Removed densehashmap references. Includes additional timers to time various callbacks. BoundaryCylinderRepulsionIn is vectorized in this version. SIMD coords, dOut calculations are generalized to accomodate upto 6 different distances. Default floatingpoint is double. Need to run a few tests to ensure float is stable.

+++ src/utility.h
@@ -35,7 +35,7 @@
 using namespace std;
 
 ///floatingpoint typedef
-typedef float floatingpoint;
+typedef double floatingpoint;
 
 //to test for zero values
 const floatingpoint ZERO_PREC = 1E-6;
